Don't call the block twice for #multi
You might want to do things in the multi block such as read a key that
you passed to `watch:`. Currently, if you do so, you will actually read
the key _twice_. The second copy of your transaction is the one that
actually gets committed.

For example, this transaction adds _two_ to "key", not one:

```
$redis.multi(watch: ["key"]) do |m|
    old_value = $redis.call_v(["GET", "key"]).to_i
    m.call_v(["SET", old_value + 1])
end
```

This patch fixes the issue by batching up the transaction's calls in our
_own_ buffer, and only invoking @node.multi after the user block is
done.