Implement RedisClient::Cluster::Command#extract_all_keys
This is required to validate that all the keys passed in to a pinned
connection are consistent with the key passed in to `#with`. Technically
speaking, we _could_ let the Redis server just reject keys that are
cross-node. However, there _is_ a good reason to perform the validation
in the client.

Redis clusters have 16,000 odd slots, but usually only a handful of
nodes. It's quite possible you might create a transaction which operates
across slots, but have it work fine because the slots just happen to
hash to the same node. However, if a resharding event then happens,
suddednly your working-fine code will break! It's nice for users to
receive feedback straight away, even on very small development setups,
that their cross-slot transactions may well not actually work later on.