Fix #pipelined concurrently using the same Redis connection
The pipelining support works by fanning out queries to the appropriate
nodes in parallel in a scatter stage, and then, during the gather stage,
detecting where redirecting is required and issuing the commands again
on the redirected-to node.

The gathering happens in parallel to the scattering, however; it's
possible for the gather thread to issue a redirection on a connection
still in use by a scatter thread, resulting in pandemonium.

I've committed a testcase with this patch which, at least on my machine,
was failing because instead of getting "OK" in response to the SET
commands, it would get a hash which was _actually_ a response to a HELLO
command issued on the same connection by mistake.

This patch fixes things by deferring the redirections until after all of
the initial pipelines have run.