Implement RedisClient::Cluster#with
This implements support for "pinning" a cluster client to a particular
keyslot. This allows the use of WATCH/MULTI/EXEC transactions on that
node without any ambiguity.

Because RedisClient also implements watch as "yield self" and ignores
all arguments, this means that you can easily write transactions that
are compatible with both cluster and non-clustered redis (provided you
use hashtags for your keys).