Pass config: to ::Node object
Currently, it takes a few elements of config (related to replication, or
slow command timeout) as its own keyword arguments, which is nice and
explicit. However, as part of the work to fix the RedisClient circuit
breakers, I need to push down some calls to the config object into the
Node class (i.e. so Node can know how to make new connections)

Thus, let's pass the config object itself into Node, and delete the
other kwargs that become redundant.