Allow node topology instances to process _updates_ to topology
At the moment, we throw away the whole ::Node structure every time we
call `CLUSTER NODES`. This has a couple of unfortunate effects:

* If one node has failed and been replaced, we wind up disconnecting
  from _all_ of the nodes, even if they're perfectly healthy. This can
  cause an unexpected connection storm on the servers.
* Because of this, we also replace the underlying RedisClient instances,
  which are responsible for maintaining circuit breakers. We actually
  therefore wind up constructing a new circuit breaker every time we
  process any topology change, and so the circuit breaker will never
  open. This, again, can lead to undesirable load on the Redis servers.

This patch is the first step towards correcting the problem. It makes
the topology class instances have a method `process_topology_update!`
which gives the the ability to maintain RedisClient instances across
topology reconfigurations. The idea is that when we call CLUSTER NODES
to find out about new nodes, we don't needlessly close and re-open
connections to nodes we already know about, instead only closing old
hosts that are gone & opening new connections to new hosts.

The implementation of this logic is the same across all of the topology
classes, so I refactored them to have a common base class (they almost
all shared the logic from the replica mixin anyway).

This logic is not acutally plumbed in anywhere yet; it's purely an
internal, behaviour preserving refactor right now because
process_topology_update! is only called from #initialize.