Implement RedisClient::Cluster#with
This PR actually implements the #with method for checking out a
connection directly to a particular Redis node. This allows you to
perform transactional commands directly against the underlying
`RedisClient` instance, so `#watch`, `#multi`, etc all work.

This does NOT yet implement any kind of validation for the commands
issued. If you issue commands involving keys not owned by the node, you
will receive `Redis::CommandError`'s. A future PR will add client-side
validation to hopefully make it possible to catch problems related to
transaction misuse earlier in development. There's a pending test for
this.

This PR also does not yet implement `RedisClient::Cluster#multi` in
terms of `#with`. That will also come in a future PR.