Use a custom subclass of RedisClient for the node clients
I tried really hard to avoid doing this, but unfortunately I don't think
we can. In order for pinning support to be safe, the underlying
RedisCluster instances we hand out from the `#with` method need to
know that they're "locked" to a particular slot, so that they can reject
cross-slot accesses that might become invalid on a cluster scale out or
in.

I had thought it would be possible to do this with middleware, but, the
middleware can't easily be accessed from `#with` to set & unset the
pinning state.