Don't attempt redirection on the actual MULTI for WATCH...MULTI
If we have a WATCH, then the MULTI _must_ exec on exactly that node; it
should not be allowed for that to be redirected to a different node,
because then the MULTI isn't on the same connection as the WATCH
anymore!

The first part of this patch fixes this by disabling redirection
handling in transaction.rb if we are in a watch. I also added a test
test_the_state_of_cluster_resharding_with_reexecuted_watch for this.

However, the second part of this patch is a lot trickier...

This change causes a different test,
test_the_state_of_cluster_resharding_with_transaction_and_watch, to
break. That test is asserting that
- if we watch something where the slot is in the middle of migrating
  between nodes,
- and all the keys we wanted to watch are on the new node,
- that the client correctly retries the transaction with ASKING against
  the new node

Disabling redirection handling obviously makes this stop working, and it
_is_ possible to handle this case correctly. We need to record whether
or not we had to issue an ASKING on the WATCH for the transaction, and
if so, pre-emptively issue an ASKING on the MULTI too. That's because
this slot is not yet actually assigned to the node we're connected to
(it's IMPORTING).

It may well not be worth it, and I'm also OK with just failing
WATCH/MULTI on slots which are currently migrating. That would imply:

- Keeping test_the_state_of_cluster_resharding_with_reexecuted_watch
- Deleting test_the_state_of_cluster_resharding_with_transaction_and_watch